% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_and_stretch.R
\name{apply_best_normalisation}
\alias{apply_best_normalisation}
\title{Apply normalisation (after applying stretch)}
\usage{
apply_best_normalisation(
  data,
  best_shifts,
  accession_data_to_transform,
  accession_data_ref
)
}
\arguments{
\item{data}{Input data (after applying stretching).}

\item{best_shifts}{Input dataframe containing information of best shifts.}

\item{accession_data_to_transform}{Accession name of data which will be transformed.}

\item{accession_data_ref}{Accession name of reference data.}
}
\value{
Normalised data.
}
\description{
\code{apply_best_normalisation} is a function to normalise by the mean and standard deviation of the compared points (after applying stretching) for each gene, in each accesion (reference data and data to transform). If the gene wasn't compared, set the expression value to NA.
}
