% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{apply_shift_to_registered_genes_only}
\alias{apply_shift_to_registered_genes_only}
\title{Apply shift for all registered genes}
\usage{
apply_shift_to_registered_genes_only(
  to_shift_df,
  best_shifts,
  model_comparison_dt,
  accession_data_to_transform,
  accession_data_ref,
  time_to_add
)
}
\arguments{
\item{to_shift_df}{Input data frame.}

\item{best_shifts}{Data frame containing information of best shift and stretch values.}

\item{model_comparison_dt}{Data frame containing information of comparison of BIC and AIC for registred and non-registered genes.}

\item{accession_data_to_transform}{Accession name of data which will be transformed.}

\item{accession_data_ref}{Accession name of reference data.}

\item{time_to_add}{Time points to be added in both reference data and data to transform after shifting and stretching.}
}
\value{
Data frame for all transformed genes for those with better BIC values.
}
\description{
\code{apply_shift_to_registered_genes_only} is a function to apply shift for all registered model based on \code{model_comparison_dt} using information from \code{best_shifts}.
}
