% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_and_stretch.R
\name{apply_stretch}
\alias{apply_stretch}
\title{Apply stretch factor}
\usage{
apply_stretch(
  data,
  best_shifts,
  accession_data_to_align = "Col0",
  accession_data_target = "Ro18",
  data_to_align_time_added = 11,
  data_target_time_added = 11
)
}
\arguments{
\item{data}{Input data.}

\item{best_shifts}{Input data frame containing information of best shifts.}

\item{accession_data_to_align}{Accession name of data which will be aligned.}

\item{accession_data_target}{Accession name of data target.}

\item{data_to_align_time_added}{Time points to be added in data to align.}

\item{data_target_time_added}{Time points to be added in data target.}
}
\value{

}
\description{
Apply stretch factor
}
