% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calcs.R
\name{calc_AIC}
\alias{calc_AIC}
\title{Calculate Akaike's ‘An Information Criterion’}
\usage{
calc_AIC(logL, num_params)
}
\arguments{
\item{logL}{Loglikelihoods value obtained by stats::logLik().}

\item{num_params}{Number of parameters in the fitted model.}
}
\value{
AIC criterion value.
}
\description{
\code{calc_AIC} is a generic function calculating Akaike's ‘An Information Criterion’ for one or several fitted model objects for which a log-likelihood value can be obtained.
}
