% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calcs.R
\name{calc_BIC}
\alias{calc_BIC}
\title{Calculate Bayesian Information Criterion}
\usage{
calc_BIC(logL, num_params, num_obs)
}
\arguments{
\item{logL}{Loglikelihoods value obtained by stats::logLik().}

\item{num_params}{Number of parameters in the fitted model.}

\item{num_obs}{Number of observations.}
}
\value{
AIC criterion value.
}
\description{
\code{calc_BIC} is a function calculating Bayesian Information Criterion, which is a special case of AIC when k = log(n), where n being the number of observations.
}
