% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calcs.R
\name{calc_score}
\alias{calc_score}
\title{Calculate mean score of observed expression data and expected value}
\usage{
calc_score(data_to_transform_expression, data_fix_expression)
}
\arguments{
\item{data_to_transform_expression}{Input expression of data_to_transform.}

\item{data_fix_expression}{Input expression of data_fix.}
}
\value{
Mean of score value.
}
\description{
\code{calc_score} is a function to calculate a mean of a score of difference of observed expression data and expected expression value, by executing score = (sum(observed-expected)**2). The mean is taken to compare the variable number of datapoints.
}
