% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_best_shifts_stretches.R
\name{calculate_all_best_shifts}
\alias{calculate_all_best_shifts}
\title{Calculate scores for all candidate shifts for all genes in data frame.}
\usage{
calculate_all_best_shifts(
  num_shifts,
  mean_df,
  stretch_factor,
  do_rescale,
  shift_extreme,
  min_num_overlapping_points,
  accession_data_to_transform = "Col0",
  accession_data_ref = "Ro18"
)
}
\arguments{
\item{num_shifts}{Number of different shifts to be considered.}

\item{mean_df}{Input data.}

\item{stretch_factor}{Current stretch factor.}

\item{do_rescale}{Apply "scale" to compared points for each shift if \code{TRUE}, use original mean expression data if \code{FALSE}.}

\item{shift_extreme}{Approximation of maximum and minimum shifts allowed.}

\item{min_num_overlapping_points}{Bound the extreme allowed shifts, such than at least this many timepoints are being compared for both accessions.}

\item{accession_data_to_transform}{Accession name of data which will be transformed.}

\item{accession_data_ref}{Accession name of reference data.}
}
\value{
\code{all_scores_df}
}
\description{
\code{calculate_all_best_shifts} is a function to get score for all shifts value after stretching with a particular stretch factor. The extremes shift values were defined by allowing 5 overlapping points for comparison.
}
