% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Calculate distance between sample data before and after registration}
\usage{
calculate_distance(results, type = c("registered", "original"))
}
\arguments{
\item{results}{Result of registration process using \code{\link{register}}.}

\item{type}{Type of comparison, determines whether to use "registered" or "original" time points. By default, "registered".}
}
\value{
This function returns a list of data frames which includes:
\item{distance_mean_df}{distance of mean expression values.}
\item{distance_scaled_mean_df}{distance of scaled mean expression (all genes).}
\item{distance_scaled_mean_df_only_nonreg}{distance of scaled mean expression (only non-registered genes).}
\item{distance_scaled_mean_df_only_reg}{distance of scaled mean expression (only registered genes).}
\item{distance_registered_df}{distance of registered & scaled mean expression (all genes).}
\item{distance_registered_df_only_reg}{distance of registered & scaled mean expression (only registered genes).}
}
\description{
Calculate distance between sample data before and after registration
}
