% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distance.R
\name{calculate_pairwise_sample_distance_main}
\alias{calculate_pairwise_sample_distance_main}
\title{Calculate pairwise sample distance}
\usage{
calculate_pairwise_sample_distance_main(df)
}
\arguments{
\item{df}{Dataframe contains the wide format expression of the two samples to be compared. Only genes which have data in both samples are considered (only relevant for registration case).}
}
\value{
Dataframe contains the distance between two samples.
}
\description{
Calculate pairwise sample distance
}
