% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_and_stretch.R
\name{compare_registered_to_unregistered_model}
\alias{compare_registered_to_unregistered_model}
\title{Comparing registered to unregistered model}
\usage{
compare_registered_to_unregistered_model(
  curr_sym,
  all_data_df,
  accession_data_to_transform,
  accession_data_ref
)
}
\arguments{
\item{curr_sym}{A gene accession.}

\item{all_data_df}{Input data.}

\item{accession_data_to_transform}{Accession name of data which will be transformed.}

\item{accession_data_ref}{Accession name of reference data.}
}
\value{
Score of AIC and BIC for both registered and unregistered models.
}
\description{
\code{compare_registered_to_unregistered_model} is a function to compare the overlapping timepoints in reference data and data to transform after best registration and without registration. Same timepoints and stretched data were used for both models.
}
