% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{get_all_data}
\alias{get_all_data}
\title{Get all data}
\usage{
get_all_data(
  filepath_data_ref,
  filepath_data_to_transform,
  filepath_id_table,
  fix_id_table_shared_colname = "CDS.model",
  fix_and_to_transform_data_shared_colname = "locus_name",
  colnames_id_table = c("CDS.model", "symbol", "locus_name"),
  colnames_wanted = NULL
)
}
\description{
Get all data
}
