% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_approximate_stretch}
\alias{get_approximate_stretch}
\title{Get approximate stretch factor}
\usage{
get_approximate_stretch(
  input_df,
  accession_data_to_transform,
  accession_data_ref
)
}
\arguments{
\item{input_df}{Input data frame contains all replicates of gene expression in each genotype at each timepoint.}

\item{accession_data_to_transform}{Accession name of data which will be transformed.}

\item{accession_data_ref}{Accession name of reference data.}
}
\description{
Get approximate stretch factor
}
