% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_and_stretch.R
\name{get_best_result}
\alias{get_best_result}
\title{Get the best result}
\usage{
get_best_result(df)
}
\arguments{
\item{df}{Input data frame containing value of applied shifts, stretches, and calculated score.}
}
\value{

}
\description{
\code{get_best_result} is a function to get best result obtained from score calculated from applied shifts and stretch factors.
}
