% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_best_shifts_stretches.R
\name{get_best_shift}
\alias{get_best_shift}
\title{Calculate the score for all shifts}
\usage{
get_best_shift(
  num_shifts = 25,
  curr_sym,
  data,
  stretch_factor,
  do_rescale,
  min_shift,
  max_shift,
  accession_data_to_transform = "Col0",
  accession_data_ref = "Ro18"
)
}
\arguments{
\item{num_shifts}{Number of different shifts to be considered.}

\item{curr_sym}{Current gene accession.}

\item{data}{Input data.}

\item{stretch_factor}{Current stretch factor.}

\item{do_rescale}{Apply "scale" to compared points for each shift if \code{TRUE}, use original mean expression data if \code{FALSE}.}

\item{min_shift}{Minimum extreme value of shift.}

\item{max_shift}{Maximum extreme value of shift.}

\item{accession_data_to_transform}{Accession name of data which will be transformed.}

\item{accession_data_ref}{Accession name of reference data.}
}
\description{
\code{get_best_shift} is used to calculate the score for all shifts (for the current gene, and current stretch_factor), and return the scores for all as a table, and the value of the optimal shift. Shift extremes are defined s.t. at least 5 points are compared.
}
