% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_best_shifts_stretches_copy.R
\name{get_best_shift_copy}
\alias{get_best_shift_copy}
\title{Calculate the score for all shifts}
\usage{
get_best_shift_copy(
  curr_sym,
  data,
  stretch_factor,
  do_rescale,
  min_shift,
  max_shift,
  testing = FALSE,
  num_shifts = 25,
  accession_data_to_align = "Col0"
)
}
\arguments{
\item{curr_sym}{Current gene accession}

\item{data}{Input data.}

\item{stretch_factor}{Stretch transformation factor wanted.}

\item{do_rescale}{Apply "scale" to compared points for each shift if TRUE, and means use original mean expression data if FALSE.}

\item{min_shift}{Minimum values of shifts.}

\item{max_shift}{Maximum values of shifts.}

\item{testing}{Make plot of shifted, and normalised gene expression if TRUE.}

\item{num_shifts}{Number of different shifts to be considered.}

\item{accession_data_to_align}{Accession name of data which will be aligned.}
}
\description{
\code{get_best_shift} is used to calculate the score for all shifts for current gene and stretch_factor, and return the scores for all as a table, and the value of the optimal shift.
}
