% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{get_best_stretch_and_shift}
\alias{get_best_stretch_and_shift}
\title{Calculate best shifts and stretches for each gene, also calculate AIC/BIC under registration or non-registration}
\usage{
get_best_stretch_and_shift(
  to.shift.df,
  all.data.df,
  stretches,
  do_rescale,
  min_num_overlapping_points,
  shift_extreme
)
}
\arguments{
\item{stretches}{Vector data of stretches.}

\item{do_rescale}{Apply "scale" to compared points for each shift if TRUE, use original mean expression data if FALSE.}

\item{min_num_overlapping_points}{Bound the extreme allowed shifts, such than at least this many timepoints are being compared for both accessions.}

\item{shift_extreme}{Approximation of maximum and minimum shifts allowed.}

\item{to_shift_df}{Input data containing mean of each time point.}

\item{all_data_df}{Input all data (without taking mean).}

\item{num_shifts}{Number of different shifts to be considered.}

\item{testing}{Showing a plot of the progress if TRUE, otherwise if FALSE.}

\item{accession_data_to_align}{Accession name of data which will be aligned.}

\item{accession_data_target}{Accession name of data target.}

\item{data_to_align_time_added}{Time points to be added in data to align.}

\item{data_target_time_added}{Time points to be added in data target.}
}
\value{
List of data frames (a) all_shifts : all the combos of stretching and shifting tried for each gene, (b) best_shifts : the best stretch and shift combo found for each gene, as well as info for scaling, and (c) model_comparison.dt : AIC / BIC scores for best registerd model found, compared to seperate model for each genes expression in the 2 accessions.
}
\description{
\code{get_best_stretch_and_shift} is a function to stretch in all stretches and calculates best shift, by comparing SUM of squares difference. For the best shift in each stretch, compares to separate models to calculate AIC/BIC under registration or no registration.
}
