% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{get_mean_and_all_exp_data}
\alias{get_mean_and_all_exp_data}
\title{Simplify reference data and data to transform into one dataframe}
\usage{
get_mean_and_all_exp_data(
  data_ref,
  data_to_transform,
  id_table,
  fix_id_table_shared_colname,
  fix_and_to_transform_data_shared_colname,
  colnames_id_table,
  colnames_wanted,
  tissue_wanted = NULL,
  curr_GoIs,
  sum_exp_data_ref = FALSE,
  accession_data_to_transform = "Col0",
  ids_data_ref_colnames = c("CDS.model", "locus_name"),
  max_expression_value_wanted = 5,
  exp_threshold = 0.5
)
}
\arguments{
\item{data_ref}{File name in working directory, path to file of reference data.}

\item{data_to_transform}{File name in working directory, path to file of data to transform.}

\item{id_table}{File name in working directory, path to file of ID table connecting both reference data and data.}

\item{fix_id_table_shared_colname}{Column names shared by both reference data and ID table.}

\item{fix_and_to_transform_data_shared_colname}{Column names shared by both reference data and data to transform.}

\item{colnames_id_table}{ID table column names.}

\item{colnames_wanted}{List of column names to keep from both reference data and data to transform.}

\item{tissue_wanted}{Name of tissue from which data will be compared.}

\item{curr_GoIs}{Gene of interest list.}

\item{sum_exp_data_ref}{If \code{TRUE} then sum all gene data. Default is \code{FALSE}.}

\item{accession_data_to_transform}{Accession name of data which will be transformed.}

\item{ids_data_ref_colnames}{Column names shared by both reference data and ID table, whose element needs to be unique.}

\item{max_expression_value_wanted}{Maximum value of expression desired.}

\item{exp_threshold}{Minimum expression threshold from which expression below the threshold will be removed.}
}
\value{
Combined data frame for both reference data and data to transform: (1) only containing mean of the expression, and (2) still contains replicate data.
}
\description{
Simplify reference data and data to transform into one dataframe
}
