% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{get_mean_data}
\alias{get_mean_data}
\title{Calculate mean expression values from all expression data with replicates}
\usage{
get_mean_data(
  exp,
  max_expression_value_wanted = 5,
  accession_data_to_transform
)
}
\arguments{
\item{exp}{Input data frame contains all replicates of gene expression in each genotype at each timepoint.}

\item{max_expression_value_wanted}{Maximum value of expression desired.}

\item{accession_data_to_transform}{Accession name of data which will be transformed.}
}
\value{
A data frame contains only mean expression data.
}
\description{
Calculate mean expression values from all expression data with replicates
}
