% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{impute_aligned_exp_values}
\alias{impute_aligned_exp_values}
\title{Setting aligned expression data and data target to be the same in a set of common time points}
\usage{
impute_aligned_exp_values(
  shifted_mean_df,
  accession_data_to_align,
  accession_data_target
)
}
\arguments{
\item{shifted_mean_df}{All registered data frame.}

\item{accession_data_to_align}{Accession name of data which will be aligned.}

\item{accession_data_target}{Accession name of data target.}
}
\value{

}
\description{
\code{impute_aligned_exp_values} is a function to impute aligned times at set of common time points in order to allow sample distance comparison to data target. this means that aligned expression data were imputed relative to data target time points. Since the original value of aligned data are not meant to be discarded, the imputed times are generated from minimum and maximum shifted time points of aligned data (not just data target time points).
}
