% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{impute_transformed_exp_values}
\alias{impute_transformed_exp_values}
\title{Setting transformed expression data and reference data to be the same in a set of common time points}
\usage{
impute_transformed_exp_values(
  shifted_mean_df,
  accession_data_to_transform,
  accession_data_ref
)
}
\arguments{
\item{shifted_mean_df}{All registered data frame.}

\item{accession_data_to_transform}{Accession name of data which will be transformed.}

\item{accession_data_ref}{Accession name of reference data.}
}
\description{
\code{impute_transformed_exp_values} is a function to impute transformed times at set of common time points in order to allow sample distance comparison to reference data. this means that transformed expression data were imputed relative to reference data time points. Since the original value of transformed data are not meant to be discarded, the imputed times are generated from minimum and maximum shifted time points of transformed data (not just reference data time points).
}
