% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpolate_data_ref_comparison_expression}
\alias{interpolate_data_ref_comparison_expression}
\title{Calculate prediction of reference data expression value}
\usage{
interpolate_data_ref_comparison_expression(data_to_transform_time, data_ref_dt)
}
\arguments{
\item{data_to_transform_time}{Input time from data to transform.}

\item{data_ref_dt}{Input reference data data frame.}
}
\value{
Expression prediction.
}
\description{
Calculate prediction of reference data expression value
}
