% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpolate_data_target_comparison_expression}
\alias{interpolate_data_target_comparison_expression}
\title{Calculate prediction of data target expression value}
\usage{
interpolate_data_target_comparison_expression(
  data_to_transform_time,
  data_target_dt
)
}
\arguments{
\item{data_to_transform_time}{Input time from data to transform.}

\item{data_target_dt}{Input data target data frame.}
}
\value{
Expression prediction.
}
\description{
Calculate prediction of data target expression value
}
