% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{load_mean_df}
\alias{load_mean_df}
\title{Simplify reference data and data to transform into one dataframe}
\usage{
load_mean_df(
  filepath_data_fix,
  filepath_data_to_transform,
  filepath_id_table,
  fix_id_table_shared_colname,
  fix_and_to_transform_data_shared_colname,
  colnames_id_table,
  colnames_wanted,
  tissue_wanted,
  curr_GoIs,
  sum_exp_data_fix = FALSE,
  accession_data_to_transform = "Col0",
  ids_data_fix_colnames = c("CDS.model", "locus_name"),
  max_mean_cpm_wanted = 5,
  exp_threshold = 0.5
)
}
\arguments{
\item{filepath_data_fix}{File name in working directory, path to file of reference data.}

\item{filepath_data_to_transform}{File name in working directory, path to file of data to transform.}

\item{filepath_id_table}{File name in working directory, path to file of ID table connecting both reference data and data.}

\item{fix_id_table_shared_colname}{Column names shared by both reference data and ID table.}

\item{fix_and_to_transform_data_shared_colname}{Column names shared by both reference data and data to transform.}

\item{colnames_id_table}{ID table column names.}

\item{colnames_wanted}{List of column names to keep from both reference data and data to transform.}

\item{tissue_wanted}{Name of tissue from which data will be compared.}

\item{curr_GoIs}{Gene of interest list.}

\item{sum_exp_data_fix}{default is FALSE. If TRUE then sum all gene data.}

\item{accession_data_to_transform}{Accession name of data which will be transformed.}

\item{ids_data_fix_colnames}{Column names shared by both reference data and ID table, whose element needs to be unique.}

\item{max_mean_cpm_wanted}{Maximum value of expression desired.}

\item{exp_threshold}{Minimum expression threshold from which expression below the threshold will be removed.}
}
\description{
Simplify reference data and data to transform into one dataframe
}
