% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heatmaps.R
\name{make_data_heatmaps}
\alias{make_data_heatmaps}
\title{Make data heatmaps}
\usage{
make_data_heatmaps(
  D.mean,
  D.scaled,
  D.registered,
  D.scaled.NR,
  D.scaled.R,
  D.registered.R
)
}
\description{
Make data heatmaps
}
