% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calcs.R
\name{min_is_compared_to_data_to_align}
\alias{min_is_compared_to_data_to_align}
\title{Get the smallest time point of data target}
\usage{
min_is_compared_to_data_to_align(data_to_align_time, data_target)
}
\arguments{
\item{data_to_align_time}{Minimum time points of candidate data to align.}

\item{data_target}{A data frame containing data target.}
}
\description{
\code{min_is_compared_to_data_to_align} is used to get the smallest time data target which is used in comparison t the data_to_align time.
}
