% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calcs.R
\name{min_is_compared_to_data_to_transform}
\alias{min_is_compared_to_data_to_transform}
\title{Get the smallest time point of reference data}
\usage{
min_is_compared_to_data_to_transform(data_to_transform_time, data_ref)
}
\arguments{
\item{data_to_transform_time}{Minimum time points of candidate data to transform.}

\item{data_ref}{A data frame containing reference data.}
}
\description{
\code{min_is_compared_to_data_to_transform} is used to get the smallest time reference data which is used in comparison t the data_to_transform time.
}
