% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calcs.R
\name{my_scale}
\alias{my_scale}
\title{Scaling value}
\usage{
my_scale(input)
}
\arguments{
\item{input}{Input data.}
}
\value{
Scaled value.
}
\description{
\code{my_scale} is a function to get a scaled value of an input by dividing the maximum value.
}
