% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot}
\alias{plot}
\alias{plot.res_greatR}
\alias{plot.dist_greatR}
\title{Visualise registration results}
\usage{
\method{plot}{res_greatR}(
  x,
  type = c("result", "original"),
  genes_list = NULL,
  show_rep_mean = FALSE,
  ncol = NULL,
  title = NULL,
  ...
)

\method{plot}{dist_greatR}(
  x,
  type = c("result", "original"),
  match_timepoints = TRUE,
  title = NULL,
  ...
)
}
\arguments{
\item{x}{Input object.
\itemize{
\item For \code{\link[=plot.res_greatR]{plot.res_greatR()}}: registration results, output of the \code{\link[=register]{register()}} registration process.
\item For \code{\link[=plot.dist_greatR]{plot.dist_greatR()}}: pairwise distances between reference and query time points, output of \code{\link[=calculate_distance]{calculate_distance()}}.
}}

\item{type}{Type of plot, whether to use registration "result" or "original" time points. By default, "result".}

\item{genes_list}{Optional vector indicating the \code{gene_id} values to be plotted.}

\item{show_rep_mean}{Whether to show \code{replicate} mean values.}

\item{ncol}{Number of columns in the plot grid. By default this is calculated automatically.}

\item{title}{Optional plot title.}

\item{...}{Arguments to be passed to methods (ignored).}

\item{match_timepoints}{If \code{TRUE}, will match query time points to reference time points.}
}
\value{
\itemize{
\item For \code{\link[=plot.res_greatR]{plot.res_greatR()}}: plot of genes of interest after registration process (\code{type = "result"}) or showing original time points (\code{type = "original"}).
\item For \code{\link[=plot.dist_greatR]{plot.dist_greatR()}}: distance heatmap of gene expression profiles over time between reference and query.
}
}
\description{
Visualise registration results
}
