% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_result.R
\name{plot_registered_gene_of_interest}
\alias{plot_registered_gene_of_interest}
\title{Plot gene of interest after registration}
\usage{
plot_registered_gene_of_interest(
  df,
  gene_accession = "all",
  title = NULL,
  ncol = NULL,
  sync_timepoints = FALSE
)
}
\arguments{
\item{df}{Data frame input after registration.}

\item{gene_accession}{List of gene accessions, default is \code{all}.}

\item{title}{Optional plot title.}

\item{ncol}{Number of columns in the plot grid. By default this is calculated automatically.}

\item{sync_timepoints}{Whether to synchronise maximum time points for each accession, by default \code{FALSE}.}
}
\value{
Plot of gene of interest after registration.
}
\description{
Plot gene of interest after registration
}
