% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_result.R
\name{plot_registered_GoIs_for_comparible_timepoints}
\alias{plot_registered_GoIs_for_comparible_timepoints}
\title{Plot gene of interest after registration}
\usage{
plot_registered_GoIs_for_comparible_timepoints(df)
}
\arguments{
\item{df}{Dataframe input after registration.}
}
\value{
Plot of gene of interest after registration.
}
\description{
Plot gene of interest after registration
}
