% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_registration_results}
\alias{plot_registration_results}
\title{Plot gene of interest after registration}
\usage{
plot_registration_results(
  results,
  type = c("registered", "original"),
  genes_list = NULL,
  title = NULL,
  ncol = NULL,
  plot_mean_data = FALSE
)
}
\arguments{
\item{results}{Registration results, output of the \code{\link{register}} registration process.}

\item{type}{Type of plot, determines whether to use "registered" or "original" time points. By default, "registered".}

\item{genes_list}{Optional vector indicating the \code{gene_id} values to be plotted.}

\item{title}{Optional plot title.}

\item{ncol}{Number of columns in the plot grid. By default this is calculated automatically.}

\item{plot_mean_data}{Whether mean data is displayed or not.}
}
\value{
Plot of genes of interest after registration process (\code{type = "registered"}) or showing original time points (\code{type = "original"}).
}
\description{
Plot gene of interest after registration
}
