% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register}
\alias{register}
\title{Register or synchronize different expression profiles}
\usage{
register(
  input,
  stretches = NA,
  shifts = NA,
  reference,
  query,
  overlapping_percent = 0.5,
  optimise_registration_parameters = FALSE,
  optimisation_config = list(num_iterations = 60)
)
}
\arguments{
\item{input}{Input data frame containing all replicates of gene expression in each genotype at each time point.}

\item{stretches}{Candidate registration stretch factors to apply to query data.}

\item{shifts}{Candidate registration shift values to apply to query data.}

\item{reference}{Accession name of reference data.}

\item{query}{Accession name of query data.}

\item{overlapping_percent}{Number of minimum overlapping time points. Shifts will be only considered if it leaves at least these many overlapping points after applying the registration function.}

\item{optimise_registration_parameters}{Whether to optimise registration parameters with Simulated Annealing. By default, \code{FALSE}.}

\item{optimisation_config}{List with optional arguments to modify the Simulated Annealing optimisation.}
}
\description{
\code{register()} is a function to register expression profiles a user
wishes to compare.
}
\examples{
\dontrun{
# Load a data frame from the sample data
all_data_df <- system.file("extdata/brapa_arabidopsis_all_replicates.csv", package = "greatR") \%>\%
  utils::read.csv()

# Running the registration
registration_results <- register(
  input = all_data_df,
  reference = "Ro18",
  query = "Col0",
  stretches = 2.3,
  shifts = 1.6
)
}
}
