% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename columns of original data}
\usage{
rename_columns(data, colnames)
}
\arguments{
\item{data}{Data frame to be renamed.}

\item{colnames}{List of column names to be renamed.}
}
\value{
Data frame with renamed columns.
}
\description{
Rename columns of data so it can be used in \code{get_mean_and_all_exp_data()}.
}
