% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{scale_all_rep_data}
\alias{scale_all_rep_data}
\title{Scaling all un-averaged data}
\usage{
scale_all_rep_data(mean_df, all_rep_data, scale_func)
}
\arguments{
\item{mean_df}{Input data containing mean of each time point.}

\item{all_rep_data}{Input all data (without taking mean).}

\item{scale_func}{Scaling method choice applied in all_rep_data. There are two options: (a) "scale" where all expression values are subtracted by mean value and divided by standard deviation and (b) "my_scale" where expression values are divided by mean values.}
}
\value{
Scaled expression data in all_rep_data.
}
\description{
\code{scale_all_rep_data} is a function to scale
}
