% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{shorten_groups}
\alias{shorten_groups}
\title{Shorten groups}
\usage{
shorten_groups(exp, accession_data_to_transform = "Col0")
}
\arguments{
\item{exp}{Expression value.}

\item{accession_data_to_transform}{Accession name of data which will be transformed.}
}
\description{
Shorten groups
}
