% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_results.R
\name{summary_model_comparison}
\alias{summary_model_comparison}
\title{Summarise registration results}
\usage{
summary_model_comparison(model_comparison)
}
\arguments{
\item{model_comparison}{Input data frame, element \code{model_comparison} of result list of \code{scale_and_register_data()}.}
}
\value{
List containing summary table, registered gene accessions, and non-registered gene accessions.
}
\description{
Summarise registration results
}
