% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_and_stretch.R
\name{apply_best_shift}
\alias{apply_best_shift}
\title{Register all expression over time using optimal shift found}
\usage{
apply_best_shift(
  data,
  best_shifts,
  accession_data_to_transform,
  accession_data_fix,
  data_to_transform_time_added,
  data_fix_time_added
)
}
\arguments{
\item{data}{Input data (all data).}

\item{best_shifts}{Input data frame containing information of best shifts.}

\item{accession_data_to_transform}{Accession name of data which will be transformed.}

\item{accession_data_fix}{Accession name of data fix.}

\item{data_to_transform_time_added}{Time points to be added in data to transform.}

\item{data_fix_time_added}{Time points to be added in data fix.}
}
\value{
The registered expression over time for each gene.
}
\description{
\code{apply_best_shift} is a function to register all unregistered expression overtime using the best/optimal shifts found.
}
