% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_and_stretch.R
\name{apply_best_shift}
\alias{apply_best_shift}
\title{Register all expression over time using optimal shift found}
\usage{
apply_best_shift(
  data,
  best_shifts,
  accession_data_to_transform,
  accession_data_ref,
  time_to_add
)
}
\arguments{
\item{data}{Input data (all data).}

\item{best_shifts}{Input data frame containing information of best shifts.}

\item{accession_data_to_transform}{Accession name of data which will be transformed.}

\item{accession_data_ref}{Accession name of reference data.}

\item{time_to_add}{Time points to be added in both reference data and data to transform after shifting and stretching.}
}
\value{
The registered expression over time for each gene.
}
\description{
\code{apply_best_shift} is a function to register all unregistered expression overtime using the best/optimal shifts found.
}
