% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_and_stretch.R
\name{apply_stretch}
\alias{apply_stretch}
\title{Apply stretch factor}
\usage{
apply_stretch(
  data,
  best_shifts,
  accession_data_to_transform = "Col0",
  accession_data_ref = "Ro18",
  data_to_transform_time_added = 11,
  data_ref_time_added = 11
)
}
\arguments{
\item{data}{Input data.}

\item{best_shifts}{Input data frame containing information of best shifts.}

\item{accession_data_to_transform}{Accession name of data which will be transformed.}

\item{accession_data_ref}{Accession name of reference data.}

\item{data_to_transform_time_added}{Time points to be added in data to transform.}

\item{data_ref_time_added}{Time points to be added in reference data.}
}
\description{
Apply stretch factor
}
