% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_and_stretch.R
\name{apply_stretch}
\alias{apply_stretch}
\title{Apply stretch factor}
\usage{
apply_stretch(
  mean_df,
  best_shifts,
  accession_data_to_align,
  accession_data_target,
  data_to_align_time_added = 14,
  data_target_time_added = 14
)
}
\arguments{
\item{mean_df}{Input data.}

\item{best_shifts}{Input dataframe containing information of best shifts.}

\item{accession_data_to_align}{Accession name of data which will be aligned.}

\item{accession_data_target}{Accession name of data target.}

\item{data_to_align_time_added}{Time points to be added in data to align.}

\item{data_target_time_added}{Time points to be added in data target.}
}
\value{

}
\description{
Apply stretch factor
}
