% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_and_stretch.R
\name{apply_stretch}
\alias{apply_stretch}
\title{Apply stretch factor}
\usage{
apply_stretch(
  data,
  best_shifts,
  accession_data_to_transform,
  accession_data_ref,
  time_to_add
)
}
\arguments{
\item{data}{Input data.}

\item{best_shifts}{Input data frame containing information of best shifts.}

\item{accession_data_to_transform}{Accession name of data which will be transformed.}

\item{accession_data_ref}{Accession name of reference data.}

\item{time_to_add}{Time points to be added in both reference data and data to transform after shifting and stretching.}
}
\description{
Apply stretch factor
}
