% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shifts_extreme.R
\name{calc_extreme_shifts}
\alias{calc_extreme_shifts}
\title{Calculate extreme shifts}
\usage{
calc_extreme_shifts(
  mean_df,
  min_num_overlapping_points,
  shift_extreme,
  accession_data_to_align = "Col0",
  accession_data_target = "Ro18"
)
}
\arguments{
\item{mean_df}{Input dataframe containing data to align and data target.}

\item{min_num_overlapping_points}{Bound the extreme allowed shifts, such than at least this many timepoints are being compared for both accessions.}

\item{shift_extreme}{Approximation of maximum and minimum shifts allowed.}

\item{accession_data_to_align}{Accession name of data which will be aligned.}

\item{accession_data_target}{Accession name of data target.}
}
\value{
minimum and maximum values of shifts
}
\description{
\code{calc_extreme_shifts} is used to calculate the minimum and maximum shifts, whilst preserving the criteria that at least min_num_overlapping_points are being compared from both accessions.
}
