% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distance.R
\name{calc_sample_distance}
\alias{calc_sample_distance}
\title{Calculate sample distance wrapper}
\usage{
calc_sample_distance(df, gene_col)
}
\arguments{
\item{df}{Dataframe contains expression data from each sample.}

\item{gene_col}{Column name of accession gene.}
}
\value{
Dataframe contains squared distance of data.
}
\description{
\code{calc_sample_distance} is a wrapper to  calculate_pairwise_sample_distance to calculate distance for all compared samples.
}
