% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_and_stretch.R
\name{calculate_all_model_comparison_stats}
\alias{calculate_all_model_comparison_stats}
\title{Wrapper of applying best shifts and compare the registered and unregistered models}
\usage{
calculate_all_model_comparison_stats(
  all_data_df,
  best_shifts,
  is_testing,
  accession_data_to_transform,
  accession_data_fix,
  data_to_transform_time_added,
  data_fix_time_added
)
}
\arguments{
\item{all_data_df}{Input all data (without taking mean).}

\item{best_shifts}{Input data frame containing information of best shifts.}

\item{is_testing}{Showing a plot of the progress if \code{TRUE}, otherwise if \code{FALSE}.}

\item{accession_data_to_transform}{Accession name of data which will be transformed.}

\item{accession_data_fix}{Accession name of data fix.}

\item{data_to_transform_time_added}{Time points to be added in data to transform.}

\item{data_fix_time_added}{Time points to be added in data fix.}
}
\value{
AIC and BIC score for registered and unregistered models.
}
\description{
\code{calculate_all_model_comparison_stats} is a wrapper function to apply best shifts and compare the registered and unregistered models using compare_registered_to_unregistered_model.
}
