% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_and_stretch.R
\name{calculate_all_model_comparison_stats}
\alias{calculate_all_model_comparison_stats}
\title{Wrapper of applying best shifts and compare the registered and unregistered models}
\usage{
calculate_all_model_comparison_stats(
  all_data_df,
  best_shifts,
  accession_data_to_transform,
  accession_data_ref,
  time_to_add
)
}
\arguments{
\item{all_data_df}{Input all data (without taking mean).}

\item{best_shifts}{Input data frame containing information of best shifts.}

\item{accession_data_to_transform}{Accession name of data which will be transformed.}

\item{accession_data_ref}{Accession name of reference data.}

\item{time_to_add}{Time points to be added in both reference data and data to transform after shifting and stretching.}
}
\value{
AIC and BIC score for registered and unregistered models.
}
\description{
\code{calculate_all_model_comparison_stats} is a wrapper function to apply best shifts and compare the registered and unregistered models using compare_registered_to_unregistered_model.
}
