% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Calculate distance between sample data before and after registration}
\usage{
calculate_distance(results)
}
\arguments{
\item{results}{Result of registration process using \code{\link{register}}.}
}
\value{
This function returns a list of data frames which includes:

\item{registered}{distance between scaled reference and query expressions using registered time points.}
\item{original}{distance between scaled reference and query expressions using original time points.}
}
\description{
Calculate distance between sample data before and after registration
}
