% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_and_stretch.R
\name{compare_registered_to_unregistered_model}
\alias{compare_registered_to_unregistered_model}
\title{Comparing registered to unregistered model}
\usage{
compare_registered_to_unregistered_model(
  curr_sym,
  all_data_df,
  is_testing = FALSE,
  accession_data_to_transform = "Col0",
  accession_data_fix = "Ro18"
)
}
\arguments{
\item{curr_sym}{A gene accession.}

\item{all_data_df}{Input data.}

\item{is_testing}{Showing a plot of the progress if \code{TRUE}, otherwise if \code{FALSE}.}

\item{accession_data_to_transform}{Accession name of data which will be transformed.}

\item{accession_data_fix}{Accession name of data fix.}
}
\value{
Score of AIC and BIC for both registered and unregistered models.
}
\description{
\code{compare_registered_to_unregistered_model} is a function to compare the overlapping timepoints in data fix and data to transform after best registration and without registration. Same timepoints and stretched data were used for both models.
}
