% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_and_stretch.R
\name{compare_registered_to_unregistered_model}
\alias{compare_registered_to_unregistered_model}
\title{Comparing registered to unregistered model}
\usage{
compare_registered_to_unregistered_model(
  curr_sym,
  all_data_df,
  is_testing = FALSE,
  accession_data_to_align = "Col0",
  accession_data_target = "Ro18"
)
}
\arguments{
\item{curr_sym}{A gene accession.}

\item{all_data_df}{Input data.}

\item{is_testing}{Showing a plot of the progress if TRUE, otherwise if FALSE.}

\item{accession_data_to_align}{Accession name of data which will be aligned.}

\item{accession_data_target}{Accession name of data target.}
}
\value{
Score of AIC and BIC for both registered and unregistered models.
}
\description{
\code{compare_registered_to_unregistered_model} is a function to compare the overlapping timepoints in data target and data to align after best registration and without registration. Same timepoints and stretched data were used for both models.
}
