% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_approximate_stretch}
\alias{get_approximate_stretch}
\title{Get approximate stretch factor}
\usage{
get_approximate_stretch(
  input_df,
  accession_data_to_transform,
  accession_data_ref
)
}
\arguments{
\item{input_df}{Input data frame contains all replicates of gene expression in each genotype at each time point.}

\item{accession_data_to_transform}{Accession name of data which will be transformed.}

\item{accession_data_ref}{Accession name of reference data.}
}
\value{
This function returns an estimation of a stretch factor for registering the data.
}
\description{
\code{get_approximate_stretch()} is a function to get a stretch factor estimation given input data. This function will take the time point ranges of both reference and query data and compare them to estimate the stretch factor.
}
