% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{get_expression_of_interest}
\alias{get_expression_of_interest}
\title{Get expression of interest}
\usage{
get_expression_of_interest(
  filepath_data_ref,
  filepath_data_to_transform,
  filepath_id_table,
  fix_id_table_shared_colname = "CDS.model",
  fix_and_to_transform_data_shared_colname = "locus_name",
  colnames_id_table = c("CDS.model", "symbol", "locus_name"),
  colnames_wanted = NULL,
  tissue_wanted,
  curr_GoIs,
  sum_exp_data_ref = FALSE,
  accession_data_to_transform = "Col0",
  ids_data_ref_colnames = c("CDS.model", "locus_name")
)
}
\description{
Get expression of interest
}
