% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shifts_extreme.R
\name{get_extreme_shifts_for_all}
\alias{get_extreme_shifts_for_all}
\title{Calculate extreme shifts after the stretch transformation}
\usage{
get_extreme_shifts_for_all(
  test,
  stretch_factor,
  min_num_overlapping_points,
  shift_extreme
)
}
\arguments{
\item{stretch_factor}{Stretch transformation factor wanted.}

\item{min_num_overlapping_points}{Bound the extreme allowed shifts, such than at least this many timepoints are being compared for both accessions.}

\item{shift_extreme}{Approximation of maximum and minimum shifts allowed.}

\item{mean_df}{Input dataframe containing data to align and data target.}

\item{accession_data_to_align}{Accession name of data which will be aligned.}

\item{accession_data_target}{Accession name of data target.}
}
\value{
minimum and maximum values of shifts
}
\description{
\code{get_extreme_shifts_for_all} is used to calculate the minimum and maximum shifts can apply to data after the stretch transformation.
}
