% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{get_mean_data}
\alias{get_mean_data}
\title{Calculate mean expression values from all expression data with replicates}
\usage{
get_mean_data(
  exp,
  expression_value_threshold = 5,
  accession_data_to_transform,
  is_data_normalised = FALSE
)
}
\arguments{
\item{exp}{Input data frame contains all replicates of gene expression in each genotype at each timepoint.}

\item{expression_value_threshold}{Expression value threshold. Remove expressions if maximum is less than the threshold. If \code{NULL} keep all data.}

\item{accession_data_to_transform}{Accession name of data which will be transformed.}

\item{is_data_normalised}{TRUE if dataset has been normalised prior to registration process.}
}
\value{
A data frame contains only mean expression data.
}
\description{
Calculate mean expression values from all expression data with replicates
}
