% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_result.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Visualise distances between samples from different time points to investigate the similarity of progression of gene expression states between species before or after registration}
\usage{
make_heatmap(
  sample_dist_df,
  title = NULL,
  axis_fontsize = NULL,
  same_min_timepoint = TRUE
)
}
\arguments{
\item{sample_dist_df}{Input data frame contains sample distance between two different species.}

\item{title}{Optional plot title.}

\item{axis_fontsize}{Font size of X and Y axes labels.}

\item{same_min_timepoint}{If \code{TRUE}, the default, takes data with the same minimum timepoint.}
}
\value{
Distance heatmap of gene expression profiles over time between two different species.
}
\description{
Visualise distances between samples from different time points to investigate the similarity of progression of gene expression states between species before or after registration
}
