% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{optimise_registration_params}
\alias{optimise_registration_params}
\title{Optimise registration parameters with Simulated Annealing}
\usage{
optimise_registration_params(
  input_df,
  initial_rescale = TRUE,
  do_rescale = FALSE,
  min_num_overlapping_points = 4,
  accession_data_to_transform,
  accession_data_ref,
  num_iterations = 100
)
}
\arguments{
\item{input_df}{TODO: Input data frame containing all replicates of gene expression in each genotype at each time point.}

\item{initial_rescale}{Scaling gene expression prior to registration if \code{TRUE}.}

\item{do_rescale}{Scaling gene expression using only overlapping time points points during registration.}

\item{min_num_overlapping_points}{Number of minimum overlapping time points. Shifts will be only considered if it leaves at least these many overlapping points after applying the registration function.}

\item{accession_data_to_transform}{Accession name of data which will be transformed.}

\item{accession_data_ref}{Accession name of reference data.}

\item{num_iterations}{Maximum number of iterations of the algorithm. Default is 100.}
}
\value{
List of results. TODO.
}
\description{
Optimise registration parameters with Simulated Annealing
}
