% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_result.R
\name{plot_registration_results}
\alias{plot_registration_results}
\title{Plot gene of interest after registration}
\usage{
plot_registration_results(
  reg_result_df,
  model_comparison_df = NULL,
  gene_accession = "first_genes",
  title = NULL,
  ncol = NULL,
  sync_timepoints = FALSE
)
}
\arguments{
\item{reg_result_df}{Data frame of registration results, output from registration process.}

\item{model_comparison_df}{Data frame of model comparison, also output from registration process.}

\item{gene_accession}{List of gene accessions, default is \code{first_genes} which will take first 25 genes.}

\item{title}{Optional plot title.}

\item{ncol}{Number of columns in the plot grid. By default this is calculated automatically.}

\item{sync_timepoints}{Whether to synchronise maximum time points for each accession, by default \code{FALSE}.}
}
\value{
Plot of gene of interest after registration process.
}
\description{
Plot gene of interest after registration
}
