% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{scale_and_register_data}
\alias{scale_and_register_data}
\title{Registering data}
\usage{
scale_and_register_data(
  mean_df,
  all_data_df,
  stretches,
  shift_extreme,
  num_shifts,
  min_num_overlapping_points,
  initial.rescale,
  do_rescale,
  testing,
  accession_data_to_align,
  accession_data_target,
  data_to_align_time_added,
  data_target_time_added
)
}
\arguments{
\item{mean_df}{Input data frame contains the mean gene expression of each gene in each genotype at each timepoint.}

\item{all_data_df}{Input data frame contains all replicates of gene expression in each genotype at each timepoint.}

\item{stretches}{Candidate registration stretch factors to apply to data to align.}

\item{shift_extreme}{The absolute maximum value which can be applied as a shift to gene expression timecourse (days).}

\item{num_shifts}{Number of shifts between minimum and maximum values of shift.}

\item{min_num_overlapping_points}{Number of minimum overlapping time points.  Shifts will be only considered}

\item{initial.rescale}{Scaling gene expression prior to registration if TRUE.}

\item{do_rescale}{Scaling gene expression using only overlapping timepoints points during registration.}

\item{testing}{Showing immediate results (inclusing plots) if TRUE.}

\item{accession_data_to_align}{Accession name of data which will be aligned.}

\item{accession_data_target}{Accession name of data target.}

\item{data_to_align_time_added}{Time points to be added in data to align.}

\item{data_target_time_added}{Time points to be added in data target.}
}
\value{
List of dataframes: (a) \code{mean_df} is unchanged by \code{scale_and_register_data()}, (b) \code{mean_df_sc} is identical to \code{mean_df}, with additional column "sc.mean.cpm", (c) \code{imputed_mean_df} is registered expression data, (d) \code{all_shifts} is a table of candidate registraions applied, and score for each, and (e) \code{model_comparison_dt} is a table comparing the optimal registration function for each gene (based on \code{all_shifts} scores) to model with no registration applied.
}
\description{
\code{scale_and_register_data} is a function to process all data. This includes scaling data before registration, finding and calculate score of optimal shifts and stretches, apply the best shifts and stretches.
}
