% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_registration}
\alias{summary_registration}
\title{Summarise registration results}
\usage{
summary_registration(results)
}
\arguments{
\item{results}{Registration results, output of the \code{register} registration process.}
}
\value{
List containing summary table, registered gene accessions, and non-registered gene accessions.
}
\description{
Summarise registration results
}
